Convert sqlite tests from ctest to gtest. All tests maintain the same functionality as the original ctest versions.
Fix GTest compliancet

Replace long MIT License preamble with concise two-line format

Apply clang-format to sqlite test files

Remove original ctest sqlite test files.

Fix compilation errors

Fix GTest compliance: Move TEST_P macros outside anonymous namespace

- Close anonymous namespace before test class definitions
- Move DbTest and DbMultiFileTest classes outside anonymous namespace
- Add typedefs for CPU_SqlitePerfDb_NONE and CPU_SqlitePerfDbMultiFile_NONE
- All TEST_P and INSTANTIATE_TEST_SUITE_P macros are now outside anonymous namespace
- Improve SetUp() comments to explicitly document what internal values are reset
- Complies with GTest development guidelines:
  https://github.com/ROCm/MIOpen/wiki/GTest-development

Format

 - Remove ParseCommandLineArgs() function (not standard GTest practice)
- Replace command-line arguments with environment variables for child processes
- Use ProcessEnvironmentMap to pass env vars to child processes
- Add constexpr configuration values (MP_THREADS_COUNT, etc.)
- Multi-process tests remain disabled by default (GTest-compliant)
- Fix LockFilePath ambiguity by moving it to anonymous namespace

Format

Fix undeclared identifier errors in DbMultiFileReadTestHelper

Address reviewer remark: remove main() and handle child processes via GTest Environment

Fix: Remove redundant unterminated conditional directive in sqlite_perfdb.cpp.

CI reported an unterminated #if defined(MIOPEN_ENABLE_MULTIPROCESS_TEST_SQLITE_PERFDB)
at line 1243. This was caused by a redundant nested #if starting at line 1278
that didn't have its own matching #endif, or rather, the one #endif at the end
was only matching the inner #if.
